requirements.txt
psycopg2-binary==2.9.3
Django==4.0.5
Pillow==9.1.1
django-durationwidget==1.0.5
djangorestframework==3.13.1
